Rename module from setting to site
