+++ saleor/dashboard/settings/urls.py
@@ -3,6 +3,6 @@ from django.conf.urls import url
 from . import views
 
 urlpatterns = [
-    url(r'^$', views.index, name='setting-index'),
-    url(r'^add/$', views.create, name='setting-create'),
+    url(r'^$', views.index, name='site-index'),
+    url(r'^add/$', views.create, name='site-create'),
 ]
